Finished DB Entities with sequelize-typescript
The data model is next in line. There is a simple test function to show that the database is now connected (postgres). This DB is not yet in production. So, there are still many setup functions to perform for the rest of the data pipeline to be on the DB.